+++ packages/components/src/box-control/index.tsx
@@ -33,6 +33,7 @@ import type {
 	BoxControlProps,
 	BoxControlValue,
 } from './types';
+import { maybeWarnDeprecated36pxSize } from '../utils/deprecated-36px-size';
 
 const defaultInputProps = {
 	min: 0,
@@ -64,6 +65,7 @@ function useUniqueId( idProp?: string ) {
  *
  *   return (
  *     <BoxControl
+ *       __next40pxDefaultSize
  *       values={ values }
  *       onChange={ setValues }
  *     />
@@ -154,6 +156,12 @@ function BoxControl( {
 		values: inputValues,
 		__next40pxDefaultSize,
 	};
+
+	maybeWarnDeprecated36pxSize( {
+		componentName: 'BoxControl',
+		__next40pxDefaultSize,
+		size: undefined,
+	} );
 
 	return (
 		<Grid